stock_transfer_system_ultra_product-READY/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
